# swagger_Timebank_v0.4.yaml
openapi: 3.0.1
info:
  title: Timebank
  description: 'This is API description for Timebank project.'
  version: 0.4.0
servers:
- url: http://localhost:5000/api/v1
...
paths:
...
  /serviceregister:
    get:
      security:
      - bearerAuth: []
      tags:
        - serviceregister
      summary: Get all services sorted and date limited (optional) from Service register
      parameters:
        - $ref: '#/components/parameters/sortOrderParam'
        - $ref: '#/components/parameters/sortFieldParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRegister'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /serviceregister/{id}:
    parameters:
      - name: id
        in: path
        description: ID of record in service register
        required: true
        schema:
          type: integer
          format: int64
    get:
      security:
        - bearerAuth: []
      tags:
      - serviceregister
      summary: Get record from Service register
      description: Get details of record with associated service and consumer
      operationId: getServiceRecord
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRegisterWithUserConsumer'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
    put:
      security:
      - bearerAuth: []
      tags:
      - serviceregister
      summary: Update record in Service register
      operationId: updateServiceInRegister
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ServiceRegisterUpdate'
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRegisterUpdate'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRegister'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
    delete:
      security:
      - bearerAuth: []
      tags:
      - serviceregister
      summary: Delete a record from Service register
      operationId: deleteServiceRegisterRecord
      responses:
        200:
          description: OK
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /serviceregister-create:
    post:
      security:
      - bearerAuth: []
      tags:
        - serviceregister
      summary: Create service record in Service register
      operationId: borrowingCreate
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ServiceRegister'
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRegister'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRegister'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
  /serviceregister/{id}/{hours}/{rating}:
    parameters:
      - name: id
        in: path
        description: ID of record in service register
        required: true
        schema:
          type: integer
          format: int64
      - name: hours
        in: path
        description: Number of hours needed to do service
        required: true
        schema:
          type: integer
          format: int64
      - name: rating
        in: path
        description: Rating
        required: true
        schema:
          type: integer
          format: int64
    put:
      security:
      - bearerAuth: []
      tags:
      - serviceregister
      summary: End of service and update record in Service register
      operationId: EndOfServiceInRegister
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRegister'
        400:
          $ref: '#/components/responses/BadParam'
        401:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        5XX:
          $ref: '#/components/responses/ServerError'
components:
  schemas:
...
    Service:
      required:
        - user_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Pokosim travnik
        estimate:
          type: integer
          format: int64
          example: 2
...
    ServiceRegister:
      required:
        - service_id
        - consumer_id
        - service_status
      type: object
      properties:
        id:
          type: integer
          format: int64
        service_id:
          type: integer
          format: int64
          example: 1
        consumer_id:
          type: integer
          format: int64
          example: 1
        hours:
          type: integer
          format: int64
        service_status:
          type: string
          enum:
            - inprogress
            - ended
          example: inprogress
        end_time:
          type: string
          format: date
        rating:
          type: integer
          format: int64
          example: 5
    ServiceRegisterUpdate:
      type: object
      properties:
        service_id:
          type: integer
          format: int64
          example: 1
        consumer_id:
          type: integer
          format: int64
          example: 1
        hours:
          type: integer
          format: int64
          example: 1
        service_status:
          type: string
          enum:
            - inprogress
            - ended
          example: inprogress
        end_time:
          type: string
          format: date
          example: 2022-05-26 19:00:00
    ServiceRegisterWithUserConsumer:
        allOf:
          - $ref: '#/components/schemas/ServiceRegister'
          - type: object
            properties:
              services:
                $ref: '#/components/schemas/Service'
  parameters:
    sortOrderParam:
      in: query
      name: sort
      required: false
      schema:
        type: string
        enum:
          - asc
          - desc
      description: Used to sort return values in ascending or descending order
    sortFieldParam:
      in: query
      name: field
      required: false
      schema:
        type: string
      description: Used to select sorting field from the model
...
  responses:
    NotFound:
      description: No entries found.
    NotAuthorized:
      description: Authorization information is missing or invalid.
    BadParam:
      description: Bad paramters.
    ServerError:
      description: Unexpected response error. Server fault.
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT Authorization header using the Bearer scheme.
      scheme: bearer
      bearerFormat: JWT